renames / readme additions
