+++ examples/knative-serving/knative/fake_informers.go
