+++ examples/knative-serving/knative/ingress_stub.go
