+++ examples/knative-serving/knative/scheme/scheme.go
